add bootstrap logging

+++ web/cgi-bin/bittrader/bootstrap.sh
@@ -11,11 +11,15 @@ export HOME=/home/`whoami`
 GIT_DIR=$HOME/git/bitquant
 LOG_DIR=$GIT_DIR/web/log
 
+if [ -e $LOG_DIR/bootstrap.done ] ; then
+rm -f $LOG_DIR/bootstrap.done
+rm -f $LOG_DIR/bootstrap.log
+fi
+
 (
 flock -x -n 200 || exit 1
 # Redirect STDERR to STDOUT
-echo "Saving to <a href='log/bootstrap.log' target='_blank'>log file</a>"
-rm -f $LOG_DIR/bootstrap.log
+echo "Saving to <a href='/cgi-bin/bittrader/log/bootstrap' target='_blank'>log file</a>"
 exec 6>&1
 exec > $LOG_DIR/bootstrap.log
 exec 2>&1
@@ -28,8 +32,9 @@ $GIT_DIR/git/bootstrap.sh
 echo "Starting up servers"
 sudo systemctl enable bitquant
 sudo systemctl start bitquant
+touch $LOG_DIR/bootstrap.done
 echo "(done)"
 exec 1>&6
 ) 200> $LOG_DIR/bootstrap.lock &
-echo "See progress in <a href='log/bootstrap.log' target='_blank'>log file</a>" 
+echo "See progress in <a href='/cgi-bin/bittrader/log/bootstrap' target='_blank'>log file</a>" 
 echo "</pre>"
